requirements.txt
mkdocs-material>=9.5
pymdown-extensions>=10.7
pyyaml>=6.0
